# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Travel Companion API
  description: |
    API contract for the travel planning application (Wave 0 scaffold).

    The recommendation engine is provider-agnostic. The AI provider is selected
    at runtime via the `AI_PROVIDER` environment variable: `rule` (rule-based,
    default) or `openai`.
  version: 0.1.0
  x-ai-provider: rule|openai
servers:
  - url: http://localhost:8000
    description: Local development server
tags:
  - name: Health
    description: Service liveness and health checks
  - name: Destinations
    description: Destination catalogue lookups
  - name: Trips
    description: Trip lifecycle management and export
  - name: Visits
    description: Visited attractions (足迹) and trip completion
  - name: Spots
    description: User-created places (自建地点) with optional photos
  - name: Recommendations
    description: Destination recommendation engine
  - name: Preferences
    description: User travel preferences
paths:
  /health:
    get:
      tags:
        - Health
      summary: Liveness probe
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  status: ok
  /api/recommend:
    post:
      tags:
        - Recommendations
      summary: Generate destination recommendations (rule-based)
      operationId: recommend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendRequest"
      responses:
        "200":
          description: Recommendations generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/recommend/ai:
    post:
      tags:
        - Recommendations
      summary: Generate destination recommendations via AI provider
      operationId: recommendAi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendRequest"
      responses:
        "200":
          description: AI-powered recommendations generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          description: AI provider unavailable (not configured or API error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/destinations:
    get:
      tags:
        - Destinations
      summary: List destination catalogue
      operationId: listDestinations
      parameters:
        - name: region
          in: query
          required: false
          description: Filter destinations by region (e.g. Asia, Europe)
          schema:
            type: string
        - name: season
          in: query
          required: false
          description: Filter destinations by recommended season
          schema:
            type: string
            enum:
              - spring
              - summer
              - autumn
              - winter
      responses:
        "200":
          description: Destination catalogue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"
  /api/trips:
    post:
      tags:
        - Trips
      summary: Create a new trip
      operationId: createTrip
      parameters:
        - $ref: "#/components/parameters/XDeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trip"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      tags:
        - Trips
      summary: List all trips
      operationId: listTrips
      parameters:
        - $ref: "#/components/parameters/XDeviceId"
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
  /api/trips/{id}:
    delete:
      tags:
        - Trips
      summary: Delete a trip by id
      operationId: deleteTrip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Trip deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{id}/complete:
    post:
      tags:
        - Trips
      summary: Mark a trip as done and record its stops as visits
      operationId: completeTrip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripCompleteIn"
      responses:
        "200":
          description: Trip marked as done (visits upserted idempotently)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/visits:
    get:
      tags:
        - Visits
      summary: List visited attractions for the current device
      operationId: listVisits
      parameters:
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/DeviceQuery"
      responses:
        "200":
          description: List of visits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"
    post:
      tags:
        - Visits
      summary: Record a visited attraction for the current device
      operationId: createVisit
      parameters:
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/DeviceQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitIn"
      responses:
        "201":
          description: Visit recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/visits/{id}:
    delete:
      tags:
        - Visits
      summary: Delete a visited attraction by id (own-device only)
      operationId: deleteVisit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/DeviceQuery"
      responses:
        "204":
          description: Visit deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/spots:
    get:
      tags:
        - Spots
      summary: List user-created places for the current device
      operationId: listSpots
      parameters:
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/DeviceQuery"
        - name: destination_id
          in: query
          required: false
          description: Filter spots linked to a destination from the catalogue
          schema:
            type: integer
      responses:
        "200":
          description: List of spots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spot"
    post:
      tags:
        - Spots
      summary: Create a user-created place for the current device
      operationId: createSpot
      parameters:
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/DeviceQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpotIn"
      responses:
        "201":
          description: Spot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spot"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/spots/photo:
    post:
      tags:
        - Spots
      summary: Upload a spot photo (jpg/png/webp, max 5MB)
      operationId: uploadSpotPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUpload"
        "400":
          description: Photo rejected (wrong type or empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Photo exceeds the 5MB limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/spots/{id}:
    delete:
      tags:
        - Spots
      summary: Delete a user-created place by id (own-device only)
      operationId: deleteSpot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/DeviceQuery"
      responses:
        "204":
          description: Spot deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{id}/export:
    get:
      tags:
        - Trips
      summary: Export a trip itinerary
      operationId: exportTrip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: format
          in: query
          required: false
          description: Export file format
          schema:
            type: string
            enum:
              - markdown
              - ics
            default: markdown
      responses:
        "200":
          description: Exported itinerary
          content:
            text/markdown:
              schema:
                type: string
            text/calendar:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/preferences:
    get:
      tags:
        - Preferences
      summary: Get current user preferences
      operationId: getPreferences
      responses:
        "200":
          description: Current preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
    put:
      tags:
        - Preferences
      summary: Upsert user preferences
      operationId: upsertPreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Preferences"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
        "422":
          $ref: "#/components/responses/ValidationError"
components:
  schemas:
    Destination:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          description: Unique destination id
        name:
          type: string
          description: Destination name
        country:
          type: string
          description: Country or region name
        region:
          type: string
          description: Geographic region (e.g. Asia, Europe)
        description:
          type: string
        best_season:
          type: string
          enum:
            - spring
            - summer
            - autumn
            - winter
          description: Recommended visiting season
        tags:
          type: array
          items:
            type: string
          description: Interest tags (e.g. food, nature, culture)
        image_url:
          type: string
          format: uri
        cost_level_1:
          type: integer
        cost_level_2:
          type: integer
        cost_level_3:
          type: integer
        cost_level_4:
          type: integer
        attractions:
          type: array
          items:
            $ref: '#/components/schemas/PlaceInfo'
        hotels:
          type: array
          items:
            $ref: '#/components/schemas/PlaceInfo'
    PlaceInfo:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        day:
          type: integer
        url:
          type: string
        booking_url:
          type: string
        phone:
          type: string
        address:
          type: string
        station:
          $ref: '#/components/schemas/StationInfo'
    StationInfo:
      type: object
      required:
        - name
        - line
      properties:
        name:
          type: string
          description: Nearest station name
        line:
          type: string
          description: Railway/bus line name
    Trip:
      type: object
      required:
        - title
        - start_date
        - end_date
      properties:
        id:
          type: integer
          readOnly: true
          description: Unique trip id (server-assigned)
        title:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        destination_id:
          type: integer
          description: Optional linked destination from the catalogue
        notes:
          type: string
        status:
          type: string
          enum:
            - planned
            - done
          default: planned
          description: Trip stage (planned until completed)
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    RecommendRequest:
      type: object
      required:
        - start_date
        - end_date
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        origin:
          type: string
          description: Departure location
        budget:
          type: number
          description: Maximum budget in local currency
        interests:
          type: array
          items:
            type: string
          description: Interest tags to prioritise
        region:
          type: string
          description: Preferred region filter
        holiday_type:
          type: string
          description: Holiday mapping hint (weekend | three_day | obon | golden_week | custom)
          enum:
            - weekend
            - three_day
            - obon
            - golden_week
            - custom
    Recommendation:
      type: object
      required:
        - destination
        - score
      properties:
        destination:
          $ref: "#/components/schemas/Destination"
        score:
          type: number
          format: float
          description: Relevance score (0-100)
        reason:
          type: string
          description: Human-readable explanation of the match
        matched_interests:
          type: array
          items:
            type: string
          description: Interests the recommendation matched on
    Preferences:
      type: object
      properties:
        origin:
          type: string
          description: Default departure location
        interests:
          type: array
          items:
            type: string
          description: Default interest tags
        budget:
          type: number
          description: Default maximum budget
        ai_provider:
          type: string
          enum:
            - rule
            - openai
          description: Preferred AI provider for /api/recommend/ai
    Visit:
      type: object
      required:
        - id
        - device_id
        - attraction_name
        - visited_at
      properties:
        id:
          type: integer
          readOnly: true
          description: Unique visit id (server-assigned)
        device_id:
          type: string
          description: Device that recorded the visit
        destination_id:
          type: integer
          description: Optional linked destination from the catalogue
        attraction_name:
          type: string
          description: Name of the visited attraction
        visited_at:
          type: string
          format: date-time
          description: When the attraction was visited
    VisitIn:
      type: object
      required:
        - attraction_name
      properties:
        destination_id:
          type: integer
          description: Optional linked destination from the catalogue
        attraction_name:
          type: string
          description: Name of the visited attraction
        visited_at:
          type: string
          format: date-time
          description: When the attraction was visited (defaults to now)
    Spot:
      type: object
      required:
        - id
        - device_id
        - name
        - lat
        - lng
        - created_at
      properties:
        id:
          type: integer
          readOnly: true
          description: Unique spot id (server-assigned)
        device_id:
          type: string
          description: Device that created the spot
        destination_id:
          type: integer
          description: Optional linked destination from the catalogue
        name:
          type: string
          description: Spot name
        lat:
          type: number
          description: Latitude
        lng:
          type: number
          description: Longitude
        description:
          type: string
          description: Optional note about the spot
        photo_url:
          type: string
          description: Public URL of an uploaded photo (e.g. /uploads/xxx.jpg)
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: When the spot was created
    SpotIn:
      type: object
      required:
        - name
        - lat
        - lng
      properties:
        name:
          type: string
          description: Spot name
        lat:
          type: number
          description: Latitude
        lng:
          type: number
          description: Longitude
        destination_id:
          type: integer
          description: Optional linked destination from the catalogue
        description:
          type: string
          description: Optional note about the spot
        photo_url:
          type: string
          description: Public URL of an uploaded photo (e.g. /uploads/xxx.jpg)
    PhotoUpload:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: Public URL of the uploaded photo (served under /uploads)
    TripCompleteIn:
      type: object
      required:
        - stops
      properties:
        stops:
          type: array
          items:
            type: string
          description: Names of attractions visited on the trip (deduplicated)
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
  parameters:
    XDeviceId:
      name: X-Device-Id
      in: header
      required: false
      description: Client device identifier used to scope visits and trips
      schema:
        type: string
    DeviceQuery:
      name: device_id
      in: query
      required: false
      description: Alternative to the X-Device-Id header for scoping visits and spots
      schema:
        type: string
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"